Japan/Parkison's/Victorina/garmin_server/garmin/fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pandasql
openpyxl
pandas
requests
python-dotenv
psycopg2
tensorflow
keras
matplotlib
jupyter
aiofiles